requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
base58==2.1.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
bitarray==1.2.2
black==21.12b0
bleach==4.1.0
blinker==1.4
Brlapi==0.8.2
Brotli==1.0.9
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
click==7.1.2
colorama==0.4.4
command-not-found==0.3
ConfigArgParse==1.5.3
coverage==6.2
cryptography==3.3.2
cupshelpers==1.0
cytoolz==0.11.2
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
discord==1.7.3
discord.py==1.7.3
distro==1.5.0
distro-info==1.0
Django==4.0.1
django-phone-field==1.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-nested-routers==0.93.4
duplicity==0.8.20
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
executing==0.8.2
fasteners==0.14.1
flake8==3.9.2
Flask==2.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
Flask-Testing==0.8.1
frozenlist==1.3.0
future==0.18.2
gevent==21.12.0
geventhttpclient==1.5.3
greenlet==1.1.2
h11==0.12.0
hexbytes==0.2.2
httplib2==0.18.1
idna==3.3
importlib-metadata==4.0.1
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
ipykernel==6.6.1
ipython==8.0.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.3
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.13.2
jupyterlab==3.2.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
keyring==23.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lockfile==0.12.2
locust==2.5.1
louis==3.18.0
lru-dict==1.1.7
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.3
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netaddr==0.8.0
netifaces==0.10.9
notebook==6.4.7
oauthlib==3.1.0
olefile==0.46
outcome==1.1.0
packaging==21.3
pandocfilters==1.5.0
paramiko==2.7.2
parsimonious==0.8.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
platformdirs==2.4.1
pluggy==0.13.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.4
psutil==5.9.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.1
py==1.11.0
pycairo==1.16.2
pycodestyle==2.7.0
pycparser==2.21
pycryptodome==3.14.0
pycups==2.0.1
pyflakes==2.3.1
Pygments==2.11.2
PyGObject==3.40.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==6.2.4
pytest-cov==3.0.0
pytest-flask==1.2.0
pytest-mock==3.6.1
python-apt==2.2.1
python-dateutil==2.8.1
python-debian===0.1.39ubuntu1
python-dotenv==0.19.2
pytz==2021.1
pyxdg==0.27
PyYAML==5.3.1
pyzmq==22.3.0
reportlab==3.5.66
requests==2.27.1
rlp==2.0.1
roundrobin==0.0.2
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
selenium==4.1.0
Send2Trash==1.8.0
simplejson==3.17.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.1.3
systemd-python==234
terminado==0.12.1
testpath==0.5.0
tinydb==4.6.1
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==4.0.1
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.8
usb-creator==0.3.7
varint==1.0.2
wadllib==1.3.5
wcwidth==0.2.5
web3==5.27.0
webencodings==0.5.1
websocket-client==1.2.3
websockets==9.1
Werkzeug==2.0.2
wsproto==1.0.0
xdg==5
xkit==0.0.0
yarl==1.7.2
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
